requirements.txt
# Core AI/ML Framework
langchain==0.3.26
langchain-core==0.3.68
langchain-community==0.3.27
langchain-groq==0.3.5
langchain-huggingface==0.3.0
langchain-ollama==0.3.4
langchain-openai==0.3.27
langchain-pinecone==0.2.8
langchain-text-splitters==0.3.8
langsmith==0.4.4

# LLM Providers
openai==1.93.0
groq==0.29.0
ollama==0.5.1
together==1.5.18

# Vector Database
pinecone==7.3.0

# Machine Learning
torch==2.7.1
transformers==4.53.0
sentence-transformers==5.0.0
huggingface-hub==0.33.1
scikit-learn==1.7.0

# Data Processing
pandas==2.3.0
numpy==2.3.1
scipy==1.16.0

# Visualization
streamlit==1.46.1
matplotlib==3.10.3
plotly==6.2.0
altair==5.5.0

# Document Processing
pypdf2==3.0.1
python-docx==1.2.0
openpyxl==3.1.5
beautifulsoup4==4.13.4
lxml==6.0.0

# Utilities
requests==2.32.4
python-dotenv
pydantic==2.11.7
tiktoken==0.9.0
tqdm==4.67.1
click==8.1.8